Change fromNullable to Opaleye version.
This one is more general, and turns out to be needed more often. If it
turns out we use the old version very often, we can add a synonym for
that with a different name.

I'd like to change the signature of maybeToNullable too to match
opaleye, but we use that in too many places.